refactor: extract giftPillService
