+++ src/main/java/com/twoCube/gifts/service/GiftPillService.java
@@ -1,0 +1,14 @@
+package com.twoCube.gifts.service;
+
+import com.twoCube.gifts.dto.PillListResponse;
+import com.twoCube.gifts.dto.PillRequest;
+import com.twoCube.members.domain.Member;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public interface GiftPillService {
+    List<PillListResponse> getPillList();
+    Long createPill(PillRequest pillRequest, Member member);
+}
